services/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
stripe==11.3.0
redis==5.2.1
httpx==0.28.1
emails==0.6
jinja2==3.1.4
pillow==11.0.0
python-slugify==8.0.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
factory-boy==3.3.1
faker==33.1.0

# Linting / formatting
ruff==0.8.4
mypy==1.13.0
